Recovered files, got npm run local synth to work.
