+++ SourcegraphCode.txt
@@ -1,0 +1,77 @@
+#!/usr/bin/env bash
+
+set -euxo pipefail
+
+EBS_VOLUME_DEVICE_NAME='/dev/sdb'
+DOCKER_DATA_ROOT='/mnt/docker-data'
+
+DOCKER_COMPOSE_VERSION='1.25.3'
+DEPLOY_SOURCEGRAPH_DOCKER_CHECKOUT='/home/ec2-user/deploy-sourcegraph-docker'
+
+# 🚨 Update these variables with the correct values from your fork!
+DEPLOY_SOURCEGRAPH_DOCKER_FORK_CLONE_URL='https://github.com/arjun123970/deploy-sourcegraph-docker.git'
+DEPLOY_SOURCEGRAPH_DOCKER_FORK_REVISION='v3.32.0'
+
+# Install git
+yum update -y
+yum install git -y
+
+# Clone Docker Compose definition
+git clone "${DEPLOY_SOURCEGRAPH_DOCKER_FORK_CLONE_URL}" "${DEPLOY_SOURCEGRAPH_DOCKER_CHECKOUT}"
+cd "${DEPLOY_SOURCEGRAPH_DOCKER_CHECKOUT}"
+git checkout "${DEPLOY_SOURCEGRAPH_DOCKER_FORK_REVISION}"
+
+# Format (if necessary) and mount EBS volume
+device_fs=$(lsblk "${EBS_VOLUME_DEVICE_NAME}" --noheadings --output fsType)
+if [ "${device_fs}" == "" ] ## only format the volume if it isn't already formatted
+then
+  mkfs -t xfs "${EBS_VOLUME_DEVICE_NAME}"
+fi
+mkdir -p "${DOCKER_DATA_ROOT}"
+mount "${EBS_VOLUME_DEVICE_NAME}" "${DOCKER_DATA_ROOT}"
+
+# Mount EBS volume on reboots
+EBS_UUID=$(blkid -s UUID -o value "${EBS_VOLUME_DEVICE_NAME}")
+echo "UUID=${EBS_UUID}  ${DOCKER_DATA_ROOT}  xfs  defaults,nofail  0  2" >> '/etc/fstab'
+umount "${DOCKER_DATA_ROOT}"
+mount -a
+
+# Install, configure, and enable Docker
+yum update -y
+amazon-linux-extras install docker
+systemctl enable --now docker
+sed -i -e 's/1024/262144/g' /etc/sysconfig/docker
+sed -i -e 's/4096/262144/g' /etc/sysconfig/docker
+usermod -a -G docker ec2-user
+
+# Install jq for scripting
+yum install -y jq
+
+# Edit Docker storage directory to mounted volume
+DOCKER_DAEMON_CONFIG_FILE='/etc/docker/daemon.json'
+
+## initialize the config file with empty json if it doesn't exist
+if [ ! -f "${DOCKER_DAEMON_CONFIG_FILE}" ]
+then
+  mkdir -p $(dirname "${DOCKER_DAEMON_CONFIG_FILE}")
+  echo '{}' > "${DOCKER_DAEMON_CONFIG_FILE}"
+fi
+
+## update Docker's 'data-root' to point to our mounted disk
+tmp_config=$(mktemp)
+trap "rm -f ${tmp_config}" EXIT
+cat "${DOCKER_DAEMON_CONFIG_FILE}" | jq --arg DATA_ROOT "${DOCKER_DATA_ROOT}" '.["data-root"]=$DATA_ROOT' > "${tmp_config}"
+cat "${tmp_config}" > "${DOCKER_DAEMON_CONFIG_FILE}"
+
+## finally, restart Docker daemon to pick up our changes
+systemctl restart --now docker
+
+# Install Docker Compose
+curl -L "https://github.com/docker/compose/releases/download/${DOCKER_COMPOSE_VERSION}/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+chmod +x /usr/local/bin/docker-compose
+curl -L "https://raw.githubusercontent.com/docker/compose/${DOCKER_COMPOSE_VERSION}/contrib/completion/bash/docker-compose" -o /etc/bash_completion.d/docker-compose
+
+# Run Sourcegraph. Restart the containers upon reboot.
+cd "${DEPLOY_SOURCEGRAPH_DOCKER_CHECKOUT}"/docker-compose
+docker-compose up -d
+
